+++ r5a/src/main/java/alma/logoot/network/p2p/interfaces/NetworkP2P.java
@@ -11,7 +11,7 @@ public interface NetworkP2P {
    */
   public void sendMessage(String message);
 
-  public int getPeerID();
+  public long getPeerID();
   
   public void setOnReceiveHandler(OnReceiveHandler handler);
 }
